% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small functions.R
\name{all_duplicates}
\alias{all_duplicates}
\title{duplication check}
\usage{
all_duplicates(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
boolean vector of length length(x).
}
\description{
Checks if a vector contains any duplicated values. Unlike duplicated(), it returns TRUE for all values that appear more than once.
duplicated() returns TRUE only for the 2nd element.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
