% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PriEco8800v2.R
\name{calc_flux_manual2}
\alias{calc_flux_manual2}
\title{PriEco8800 Flux calculator}
\usage{
calc_flux_manual2(
  dataset,
  height_list = rep(NA, 25),
  wt_list = rep(NA, 25),
  V_head = 6.5 * (3.85/2)^2 * pi,
  V_sys = 190,
  V_delta = 0,
  start_cutoff = 60,
  end_cutoff = 450
)
}
\arguments{
\item{dataset}{Flux measurement data for slope calculation as loaded by load_pri8800_data.}

\item{wt_list}{List of soil sample weights in each vessel in grams.  Vector of length 25 required, corresponding to sample positions.}

\item{V_head}{Air volume of the plastic cylinder heads in mL. Default is 6.5 cm x (3.85 cm)^2 x pi = 75.67 mL}

\item{V_sys}{System volume (Tubing etc). Default is 182 mL + 8 mL = 190 mL.}

\item{V_delta}{Additional volume tuning argument in mL. Default = 0.}

\item{start_cutoff}{Data is used for evaluation from this time (sec) after after start of recording. Default 60s}

\item{end_cutoff}{Data is used for evaluation to this time (sec) after after start of recording. Default 450s}

\item{heigth_list}{Sample height in the cylinder in cm}
}
\description{
Manual flux calculation for PriEco8800 incubator.
}
\note{
optimise potentially: Function for individual meas, in main function, group by run,round,sample; summarise
}
