% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_aggregate.R
\name{cm_aggregate}
\alias{cm_aggregate}
\title{cm-wise aggregation of soil data}
\usage{
cm_aggregate(
  dataset,
  depth_top_col,
  depth_bottom_col,
  aggregate_list,
  group_list,
  res_out = 0.1
)
}
\arguments{
\item{dataset}{An R tibble or data.frame}

\item{depth_top_col}{Upper sampling depth column name}

\item{depth_bottom_col}{Lower sampling depth column name}

\item{aggregate_list}{Columns to be aggregated. Can be character vector or string. Selected column must be numeric}

\item{group_list}{Optional grouping variables (e.g., sites, sampling points, soil cores etc.)}

\item{res_out}{Aggregate depth increments in cm}
}
\description{
Dataset containing sampling depth (upper+lower) and numeric values can be aggregated to equal depth increments.
}
