% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small functions.R
\name{colorblind_safe_colors}
\alias{colorblind_safe_colors}
\title{get ggthemes discrete colorblind-safe colors}
\usage{
colorblind_safe_colors()
}
\value{
8 discrete colors as hex-string vector
}
\description{
https://jrnold.github.io/ggthemes/reference/colorblind.html
}
