% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilStock.R
\name{compute_TOCstock}
\alias{compute_TOCstock}
\title{UNSTABLE Stock calculation from soil profile}
\usage{
compute_TOCstock(
  data,
  depth_from = 0,
  depth_to = 0.3,
  depth_top_var = "Depth_top",
  depth_bottom_var = "Depth_bottom",
  stock_var = "TOCstock",
  grouping_vars = c("site_id", "Profile")
)
}
\arguments{
\item{data}{Soil profile dataset including at least a Depth top and bottom, a stock variable}

\item{depth_from}{Upper boundary in m. Defualt 0}

\item{depth_to}{Lower boundary in m. Default .3}

\item{stock_var}{Stock variable name as character string}
}
\description{
Function to compute weighted TOCstock for a given depth limit (in meters)
}
