% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small functions.R
\name{conf_95}
\alias{conf_95}
\title{confidence band calculator}
\usage{
conf_95(pred, obs = mean(pred))
}
\arguments{
\item{pred}{vector with (predicted) values}

\item{obs}{vector with reference values. Default is mean(pred)}
}
\description{
calculates (linear) 95 \% confidence band around values or for obs vs. pred
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
