% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_adjusted.R
\name{evaluate_model_adjusted}
\alias{evaluate_model_adjusted}
\title{Calculate a range of evaluation metrics}
\usage{
evaluate_model_adjusted(data, obs = "obs", pred = "pred")
}
\arguments{
\item{data}{Tibble, data.frame containing at least 2 columns with numeric data (observations, predicitions)}

\item{obs}{String for selectingobservation column}

\item{pred}{String for selecting prediction column}
}
\description{
This function is based on simplerspec::evaluate_model by
Baumann P (2020). _simplerspec: Soil and plant spectroscopic model building and prediction_. R package version
0.1.0.9001, <https://github.com/philipp-baumann/simplerspec>.
Several metrics have been added to the function (e.g. R2 and linsccc)
}
