% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model handlR.R
\name{evaluate_model_batch}
\alias{evaluate_model_batch}
\title{model evaluation (BDF-SSL)}
\usage{
evaluate_model_batch(
  root_dir = "C:/Users/adam/Documents",
  model_folder = "/GitHub/R_main/models/2024 models/PLS_75_25_var_dependent_preProcess",
  model_type_pattern = "pls_",
  new_eval = F,
  verbose_iter = T,
  path_evaluate_model_adjusted =
    "/GitHub/BDF/BDF-SSL/3_r_scripts/evaluate_model_adjusted.R"
)
}
\arguments{
\item{root_dir}{Root directory to Project}

\item{model_folder}{Folder containing model objects}

\item{model_type_pattern}{prefix of model object names}

\item{new_eval}{recalculate evaluation stats with evaluate_model_adjusted.R (orginal statistics will still be saved seperately)}

\item{verbose_iter}{Print progress updates}

\item{path_evaluate_model_adjusted}{Internal argument for sourcing required functions}
}
\description{
Batch model runs evaluation aggregation
Currently only supports testset data (no gt300)
}
\note{
evaluate_model_adjusted source path is legacy. Change in future
}
