% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model handlR.R
\name{get_best}
\alias{get_best}
\title{get best model for variable based on metric}
\usage{
get_best(
  model_eval,
  type_ = "cubist",
  prefix = paste0(type_, "_"),
  variable_ = "CORG",
  metric = "test.rmse",
  maximise = F
)
}
\arguments{
\item{model_eval}{model evaluation object created with evaluate_model_batch()}

\item{type_}{Model type, i.e., "cubist", "pls", "svmLinear"...}

\item{prefix}{Model names prefix, usually defines model type. Default is type_}

\item{variable_}{Variable for which the best model is to be searched}

\item{metric}{evaluation statistic which is used to select best model}

\item{maximise}{Should the model with the highest metric be selected? E.g., for R2 set to True. Default is False.}
}
\description{
From a folder containing model object created with scripts used in the BDF-SSL framework, select the best model candidate based on a evaluation metric for a given variable.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\note{
Not yet generalised
}
