% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small functions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{ggplot default color emulator}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors to generate}
}
\description{
Emulate ggplot color scale
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
