% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PriEco8800v2.R
\name{load_pri8800_data}
\alias{load_pri8800_data}
\title{load Pri8800 data}
\usage{
load_pri8800_data(
  folder,
  contains_folders = T,
  calculate_flux_manual = F,
  V_ch_list = c(NA),
  A_ch = NA,
  wt_list = c(NA)
)
}
\arguments{
\item{folder}{Path to folder containing measurement files.}

\item{contains_folders}{Boolean. Recursive read in or not.}

\item{calculate_flux_manual}{Boolean. Should flux be re-calculated from slope data? This might be slow.}

\item{V_ch_list}{Total system volume (accumulation volume) of each incubation vessel. If NA defaults to 1. cubic meter}

\item{A_ch}{Default base area - For compatibility with field chambers. If NA defaults to 1. square meter}

\item{wt_list}{List of soil sample weights in each vessel. grams}
}
\description{
load different files created by Pri8800 (both inidvidual slopes and final output files).
}
