% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small functions.R
\name{none}
\alias{none}
\title{none}
\usage{
none(x)
}
\arguments{
\item{x}{Input}
}
\value{
x
}
\description{
Function that does absolutely nothing and returns data as is.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
