% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model handlR.R
\name{predict_variable}
\alias{predict_variable}
\title{predict variable (BDF-SSL framework)}
\usage{
predict_variable(
  target_data,
  model_folder,
  eval_model_folder = model_folder,
  prefix = "cubist_",
  variable_ = "CORG",
  metric = "test.rmse",
  maximise = F,
  restrict = T,
  diss_limit = 2.5,
  manual = NA
)
}
\arguments{
\item{prefix}{Model names prefix, usually defines model type}

\item{metric}{evaluation statistic which is used to select best model}

\item{maximise}{Should the model with the highest metric be selected? E.g., for R2 set to True. Default is False.}

\item{taget_data}{Nested tibble with spc data at different pre-processing steps}

\item{variable}{Variable for which the best model is to be searched}

\item{restrict_spc}{If True only regard models for which spc preprocessing sets are available. If False, search in all models for best candidate and
return ERROR when spc set for best candidate is not available}
}
\description{
Selection of best model for agiven variable based on evaluation metric from a set of models and prediction for target spectra
}
