% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model handlR.R
\name{predict_variable2}
\alias{predict_variable2}
\title{predict variable (generalised)}
\usage{
predict_variable2(
  target_data,
  model_folder,
  eval_model_folder = model_folder,
  prefix = "cubist_",
  variable_ = "CORG",
  metric = "test.rmse",
  maximise = F,
  restrict = T,
  diss_limit = 2.5,
  manual = NULL,
  model_obj_X_name = "spc_set",
  manual_set = "spc",
  X_scaling = 230.2585
)
}
\arguments{
\item{model_folder}{Folder containing trained models as serialised (rds) R objects.}

\item{eval_model_folder}{Location of serialised (rds) evaluation object}

\item{prefix}{Model names prefix, usually defines model type}

\item{metric}{evaluation statistic which is used to select best model.}

\item{maximise}{Should the model with the highest metric be selected? E.g., for R2 set to True. Default is False.}

\item{restrict}{If True only regard models for which spc preprocessing sets are available. If False, search in all models for best candidate and
return ERROR when spc set for best candidate is not available}

\item{diss_limit}{Mahalanobis distance limit for outlier flagging}

\item{manual}{Select model manually. \code{prefix}, \code{variable}, \code{metric}, \code{maximise}, \code{eval_model_folder} are disregarded.}

\item{model_obj_X_name}{Adjust name of spc set identifier in model$documentation. Default is spc_set. If NULL, uses manual_set}

\item{manual_set}{Manually assign spc_set for compatability}

\item{X_scaling}{Scaling factor for spectra. X = X_raw / X_scaling. I.e., for conversion of Lambda365-abs to Spctrolyzer-abs... log10 -> ln conversion and cm-1 -> m-1. Default is set to 230.2585}

\item{taget_data}{Nested tibble with spc data at different pre-processing steps, matching the structure of the calibration data used for modelling}

\item{variable}{Variable for which the best model is to be searched}
}
\description{
Selection of best model for agiven variable based on evaluation metric from a set of models and prediction for target spectra
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Wrapper for predicting variables with pre-trained caret train-objects from new spectra or MBL.
}
