% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soliTOC_readR.R
\name{pull_set_soliTOC}
\alias{pull_set_soliTOC}
\title{pull_set_soliTOC}
\usage{
pull_set_soliTOC(
  soliTOC_file,
  fix_cols = T,
  ID_col_set = "Name",
  set_id = c("LA", "LQ", "LR", "LP", "LG"),
  ID_col_std = "Name",
  std_id = "caco3",
  Memo_col_set = "Memo",
  set_memo = c("GT300"),
  std_value_col = "TC  [\%]",
  actual = 12,
  keep_batch = F,
  omit_check_cols = F
)
}
\arguments{
\item{soliTOC_file}{Excel file containing raw export from soliTOC Elementar software. Check for proper column naming, as Date Time is typically read as 2 columns, bt only one column name. This shifts all subsequent col names to the left. Fix manually.}

\item{ID_col_set}{Which column contains the identifier for the sample selection? Default is "Name"}

\item{set_id}{String or vector of strings which represent the unique identifier(s) that are searched for in `ID_col_set` by str_detect}

\item{ID_col_std}{Which columns contain the identifier for the standard selection? Default is "Name"}

\item{std_id}{String or vector of strings which represent the unique identifier(s) that are searched for in `ID_col_std` by str_detect. Default is "caco3".}

\item{Memo_col_set}{Which column contains the additional Memo identifier?}

\item{set_memo}{Optional string or vector to search for in Memo_col_set. Optional - to omit this subset layer set to NA.}

\item{std_value_col}{Which column containes the measurements}

\item{actual}{True value (concentration). Default is 12 (wt-%) TC for caco3.}

\item{keep_batch}{Should the column containing aggregation information be kept? Default=F}

\item{omit_check_cols}{Override checking of columns integrity. Default=F.}
}
\value{
A tibble or data.frame containing the dayfactor corrected C-fractions for the samples selected by detection of  ID_col
}
\description{
Function for the correction and aggregation of SoliTOC measurements. Loads a raw excel file as exported from soliTOC software. Calculates and applies dayfactor correction based on standard measurements. Select and aggregate samples based on identifer string(s).
}
