% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small functions.R
\name{R2}
\alias{R2}
\title{degree of determination}
\usage{
R2(obs, pred)
}
\arguments{
\item{obs}{Observations, reference values}

\item{pred}{Predicted values}
}
\description{
degree of determination
}
