% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPangaea.R
\name{readPangaea}
\alias{readPangaea}
\title{read Pangaea delimited text data with header}
\usage{
readPangaea(
  path,
  dec_sep = ".",
  col_sep = "\\t",
  skiplines = "auto",
  shortnames = "T"
)
}
\arguments{
\item{path}{Character string with path to Pangaea .tab-file}

\item{dec_sep}{decimal separator (default=".")}

\item{col_sep}{column separator (default="\t")}

\item{skiplines}{description lines to be skipped. Can be assinged manual by setting integer number, or by default is "auto",
looking for the end of the description marked by "*/"}

\item{shortnames}{SHould only abbreviations be returned as colnames or full descriptive titles? If True (default) only short names (abbreviations) are returned /used.
Cropping first ".." of colnames.}
}
\description{
read Pangaea delimited text data with header
}
