% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMACH_v3_readR.R
\name{SEMACHv3_plot_raw}
\alias{SEMACHv3_plot_raw}
\title{SEMACHv3 raw data plotter}
\usage{
SEMACHv3_plot_raw(
  dataset,
  variables = "CO2.SCD30",
  cutoff_start = 0,
  cutoff_end = 10800,
  rm_bad_SCD30 = T,
  rm_bad_GMP252 = T,
  leg_pos = "top"
)
}
\arguments{
\item{dataset}{Raw SEMACHv3 measurement file or glued set of measurement files (i.e., bind_rows()) loaded by \code{load_SEMACHv3_raw}, or equivalent.}

\item{variables}{Variables that are to be displayed. Can be single character or vector of character. If not available in dataset, warning will be displayed and entry is omitted. Default=c("CO2.SCD30")}

\item{cutoff_start}{Use data from when after measurement start (Default=0).}

\item{cutoff_end}{Use data to when after measurement start (Default=10800). If measurement duration is smaller, all available data is used.}

\item{rm_bad_SCD30}{Should bad SCD30 measurements be omitted from the data (negative or 0 values). If SCD30 is not plotted, this will be disregarded. Default=T}

\item{leg_pos}{Argument handed to theme(legend.positon=leg_pos). Default="top". Set "none" to draw no legend. Details see \code{ggplot2::theme()}}

\item{rm_bad_GMP252n}{Should bad GMP252 measurements be omitted from the data (negative or 0 values). If GMP252 is not plotted, this will be disregarded. Default=T}
}
\description{
Display raw measurement data loaded by \code{load_SEMACHv3_raw}, or equivalent. Designed primarily for displaying CO2 measurement data, but can also be used for other data. Plots facet for each varaible.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
