% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMACH_v3_readR.R
\name{SEMACHv3_plotRes}
\alias{SEMACHv3_plotRes}
\title{SEMACHv3 plot results}
\usage{
SEMACHv3_plotRes(pathlist = NULL, dataset = NULL, type = "R")
}
\arguments{
\item{pathlist}{Character vector containing path(s) to csv file(s) containing saved SEMACHv3 results (python SEMACHv3.exe output).
Folder structure described in \code{SEMACHv3_batch_evaluation}needs to be followed to work properly.}

\item{dataset}{A results tibble returned by \code{SEMACHv3_batch_evaluation}}

\item{type}{For compatability of python SEMACHv3.exe (=P), and R results (=R, default)}
}
\description{
Function for displaying eval_SEMACHv3 result files
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
