% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soliTOC_readR.R
\name{soliTOC_remove_duplicates}
\alias{soliTOC_remove_duplicates}
\title{duplicate filtering}
\usage{
soliTOC_remove_duplicates(
  dataset,
  measurement_method = "DIN19539",
  reference = T,
  method = "best_MAE",
  measurement_method.col_name = "Methode",
  date.col_name = "Datum",
  time.col_name = "Zeit",
  name.col_name = "Name",
  reference.col_name = "CORG",
  measured.col_name = "TOC"
)
}
\arguments{
\item{dataset}{soliTOC data.frame or tibble that has already undergone get_dayfactors.
Using raw soliTOC data could lead to loss of e.g., standard data, as they would be flagged as duplicates.}

\item{measurement_method}{soliTOC program name used (e.g. Default="DIN19539").}

\item{reference}{Logical. Is reference data provided? If TRUE, a column labeled CORG containing SOC data needs to be attached
to the data.frame. This will be compared with the measured TOC to determine best match based on MAE.}

\item{method}{duplicate resolving method.
"best_MAE" for a), only available if reference=T.
"mean" for b)
"latest" for c)}

\item{.col_name}{several arguments to set column names. Default are German names.
Adjust when using Elementar soliTOC software in different language.}
}
\description{
soliTOC_remove_duplicates
Function that removes repeated samples from soliTOC file based on
a) lowest MAE to reference data for SOC (named CORG)
b) averageing duplicates
c) selecting the latest measured replicate
}
