% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrolyzer_load.R
\name{Spectro_load_fingerprint}
\alias{Spectro_load_fingerprint}
\title{Spectrolyzer fingerprint loader}
\usage{
Spectro_load_fingerprint(
  directory,
  wavelengths = seq(200, 750, 2.5),
  sel_wavelengths = seq(200, 722.5, 2.5),
  id = T
)
}
\arguments{
\item{directory}{Path to unzipped fingerprint-folder.}

\item{wavelengths}{Vector of wavelengths. Column names of measured wavelengths. default is 200-750 nm in 2.5 nm steps}

\item{sel_wavelengths}{Vector of wavelengths. Range of wavelengths that shall be returned (<= range of `wavelengths`)}

\item{id}{Boolean. Should serial No be added as column to the tibble (for ID-ing the spectrolyzer later). Default=T. For addition of id-column, the metadata-file must be present in `directory`}
}
\value{
A nested tibble with the columns Date_Time (POSIXct), spc with spc$`sel_wavelengths` (absorbances at wavelength) and, if `id`==TRUE, the Serial_No, containing the spectrolyzer id.
}
\description{
NOTE: Removes flagged spectra. Generalised fingerprint load-in for spectrolyzer data.
`r lifecycle::badge("experimental")`
}
