% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectrolyzer_load.R
\name{Spectro_load_parameter}
\alias{Spectro_load_parameter}
\title{Spectrolyzer parameter loader}
\usage{
Spectro_load_parameter(directory, parameters = c("Temp"), id = T)
}
\arguments{
\item{directory}{Path to unzipped parameter-folder.}

\item{id}{Boolean. Should serial No be added as column to the tibble (for ID-ing the spectrolyzer later). Default=T. For addition of id-column, the metadata-file must be present in \code{directory}}
}
\value{
A tibble with the columns Date_Time (POSIXct), \code{parameters} (additionally measured variables) and, if \code{id}==TRUE, the Serial_No, containing the spectrolyzer id.
}
\description{
NOTE: Removes flagged spectra. Generalised parameter load-in for spectrolyzer data.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
