% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small functions.R
\name{summarise_metrics}
\alias{summarise_metrics}
\title{basic statistics summary}
\usage{
summarise_metrics(dataset, grouping_variables = NULL, variables = NULL)
}
\arguments{
\item{dataset}{tibble, data.frame with at least one numeric value}

\item{variables}{String or vector of strings of variables that should be summarised.}

\item{grouping_variable}{Optional. String or vector of strings describing the columns that should be used for grouping}
}
\value{
summary_df: Tibble with column varaible, if defined grouping_variable and corresponding summary statistics.
}
\description{
calculates set of basic metrics (min,max, quartiles, mean...)
}
