% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_training.R
\name{train_sample_size}
\alias{train_sample_size}
\title{Sample size effect on model accuracy estimator}
\usage{
train_sample_size(
  data,
  Xr = "spc_sg_snv_rs4",
  Yr = "TOC",
  trans = none,
  trans_rev = none,
  train_ratio = 0.7,
  kmeans_pc = 0.99,
  min_samples = 30,
  sample_step = 10,
  method = "glm",
  trControl = trainControl(method = "cv", number = 10),
  tuneGrid = NULL,
  save_all = F,
  save_models = F,
  output_folder = "models_folder",
  return_all = T,
  seed = 123,
  ...
)
}
\arguments{
\item{data}{Dataset for model calibration. Can be tibble or dataframe. Must contain column Yr and column or nested tibble/data.frame/matrix Xr.}

\item{train_ratio}{0.7, train / test split. 0...1 - fraction of data used for training}

\item{min_samples}{lowest no of samples for training.}

\item{sample_step}{steps used to increment training sample size from min_samples to nrow(data)*train_ratio}

\item{trControl}{model fitting parameters handed to caret::train}

\item{output_folder}{folder path in which models will be saved if save = TRUE. If dir.exists(output_folder) = FALSE, folder will be created.}

\item{...}{additional arguments passed to caret::train}

\item{Xr="spc_sg_snv_rs4"}{Predictor variable(s). Vector column or nested tibble/data.frame/matrix.Will be converted into nested matrix.}

\item{Yr="TOC"}{Target variable. Currently only single column supported.}

\item{trans=none, }{Transformation applied to Yr by match.fun(trans)(Yr). Set to none for no operation.}

\item{trans_rev=none, }{Reverse operation to trans. Appplied by match.fun(trans_rev)(Yr). Set to none for no operation.}

\item{kmeans_pc=.99, }{cumvar of kmeans pc decomposition. Ignored when Xr<3. (kmeans applied directly on Xr)}

\item{method="glm"}{model type (handed to caret::train). Make sure to install/load required packages. (not yet automated)}

\item{tuneGrid=NULL}{tuning parameter grid handed to caret::train}

\item{save=F}{should individual models be saved?}

\item{return_all=T}{return to console}

\item{seed=123}{seed for initial train / test split for reproduceability}
}
\value{
If return_all = TRUE, returns object containing a list 'models' with all calibrated models (caret::train object with additional details appended to object$documentation)
and val_stats containing evaluation metrics. !!! CURRENTLY USES LOCAL FUNCTION from evaluate_model_adjusted.R
}
\description{
Splits dataset into train/test set and runs multiple calibrations using increasing number of observartions from training data.
Model accuracy is calculated for each created model using the hold-out test set.
Can be used to assess effect of sample size on model performance. Should support most (numeric) regression models implemented in caret (no guarantees though - not tested)
}
\note{
In future, extract kmeans sampling function as standalone
}
