% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_training.R
\name{val_train_size}
\alias{val_train_size}
\title{val_train_size}
\usage{
val_train_size(models, trans_rev = NULL)
}
\arguments{
\item{models}{Models list object created by train_sample_size.}

\item{trans_rev}{Optional function to reverse transformation. 'none' will do nothing.
If not specified, function will catch trans_rev from models$...$documentation$trans_rev}
}
\description{
validation of model list output created by train_sample_size()
}
\note{
install additional packages for specific model types manually
}
