% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small functions.R
\name{wavenumber_wavelength}
\alias{wavenumber_wavelength}
\title{wavenumber-wavelength converter}
\usage{
wavenumber_wavelength(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
converted to respective other unit.
}
\description{
Wavenumber to wavelength conversion. Works both ways (wn-wl, wl-wn).
}
