% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PriEco8800v2.R
\name{calc_flux_manual}
\alias{calc_flux_manual}
\title{PriEco8800 Flux calculator}
\usage{
calc_flux_manual(dataset, V_ch_list = c(NA), A_ch = NA, wt_list = c(NA))
}
\arguments{
\item{dataset}{Flux measurement data for slope calculation as loaded by load_pri8800_data.}

\item{V_ch_list}{Total system volume (accumulation volume) of each incubation vessel. If NA defaults to 1. cubic meter}

\item{A_ch}{Default base area - For compatibility with field chambers. If NA defaults to 1. square meter}

\item{wt_list}{List of soil sample weights in each vessel. grams}
}
\description{
flux calculation for PriEco8800 incubator. This function might be unstable. Use at own risk.
}
\note{
optimise potentially: Function for individual meas, in main function, group by run,round,sample; summarise
}
