% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab_ID_generator.R
\name{label_creator}
\alias{label_creator}
\title{label_creator}
\usage{
label_creator(dataset, origin = "X", digits = 4)
}
\arguments{
\item{dataset}{Either vector or data.frame / tibble structure. length(dataset) or nrwo(dataset) ID labels will be created respectively.}

\item{origin}{Set an initial string which will be identical for all ID.}

\item{digits}{Total number of digits of the ID, including origin.}
}
\value{
Vector of ID strings corresponding to each entry (vector) or row (tibble) of dataset.
}
\description{
label_creator
}
