% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PriEco8800v2.R
\name{load_pri8800_data}
\alias{load_pri8800_data}
\title{load Pri8800 data}
\usage{
load_pri8800_data(
  folder,
  contains_folders = T,
  calculate_flux_manual = F,
  height_list = rep(NA, 25),
  wt_list = rep(NA, 25),
  V_head = 6.5 * (3.85/2)^2 * pi,
  V_sys = 190,
  V_delta = 0,
  start_cutoff = 60,
  end_cutoff = 450
)
}
\arguments{
\item{folder}{Path to folder containing measurement files.}

\item{contains_folders}{Boolean. Recursive read in or not.}

\item{calculate_flux_manual}{Boolean. Should flux be re-calculated from slope data? This might be slow.}

\item{wt_list}{List of soil sample weights in each vessel in grams.  Vector of length 25 required, corresponding to sample positions.}

\item{V_head}{Air volume of the plastic cylinder heads in mL. Default is 6.5 cm x (3.85 cm)^2 x pi = 75.67 mL}

\item{V_sys}{System volume (Tubing etc). Default is 182 mL + 8 mL = 190 mL.}

\item{V_delta}{Additional volume tuning argument in mL. Default = 0.}

\item{start_cutoff}{Data is used for evaluation from this time (sec) after after start of recording. Default 60s}

\item{end_cutoff}{Data is used for evaluation to this time (sec) after after start of recording. Default 450s}

\item{heigth_list}{Sample height in the cylinder in cm}
}
\description{
load different files created by Pri8800 (both inidvidual slopes and final output files).
}
