% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMACH_v3_readR.R
\name{load_SEMACHv3_file}
\alias{load_SEMACHv3_file}
\title{load_SEMACHv3_file}
\usage{
load_SEMACHv3_file(path, version = "v3.2")
}
\arguments{
\item{path}{A file path to a SEMACHv3 .csv measurement file.}

\item{version}{SEMACH version for compatability with header read in. Default v3.2 (most recent systems)}
}
\value{
Formatted raw SEMACHv3 measurement file
}
\description{
A function for loading and formatting SEMACHv3 csv measurement files.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
