% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMACH_v3_readR.R
\name{load_SEMACHv3_raw}
\alias{load_SEMACHv3_raw}
\title{laod_SEMACHv3_raw}
\usage{
load_SEMACHv3_raw(path)
}
\arguments{
\item{path}{A file path to a SEMACHv3 .csv measurement file.}
}
\value{
Formatted raw SEMACHv3 measurement files as a tibble. File name is used as ID.
}
\description{
A function to load one or multiple SEMACHv3 raw files and add them to a tibble.
}
