% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model handlR.R
\name{plot_spc}
\alias{plot_spc}
\title{plot spectra}
\usage{
plot_spc(spc, alpha = 0.1, xlab = "wavelength [nm]", ylab = "Absorbance")
}
\arguments{
\item{spc}{A tibble or data.frame with numeric colnames (i.e. `200`, `205` ...)}

\item{alpha}{Set line transparancy (lower for larger n recommended). Default = 0.1}

\item{xlab}{y-Label, Default "Absorbance"}
}
\description{
A rudimentray function for displaying spc.
}
\details{
Simple lineplot of spc is returned.
}
