% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small functions.R
\name{RMSE}
\alias{RMSE}
\title{root mean squared error}
\usage{
RMSE(obs, pred)
}
\arguments{
\item{obs}{Observations, reference values}

\item{pred}{Predicted values}
}
\description{
Root mean squared error. Nas are removed.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
